Added main menu and pause menu
Currently, the settings and bookmarks buttons don't do anything. But the explorer can be paused to open the pause menu, and "new" on the main menu instantiates a new map with default parameters.

Bookmarks are basically saves, and will later allow you to save your place in your exploration of the map.